+++ spec/nanook/wallet_spec.rb
@@ -474,6 +474,10 @@ RSpec.describe Nanook::Wallet do
             \"142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\": {
                  \"amount\": \"6000000000000000000000000000000\",
                  \"source\": \"xrb_3dcfozsmekr1tr9skf1oa5wbgmxt81qepfdnt7zicq5x3hk65fg4fqj58mbr\"
+            },
+            \"242A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\": {
+                 \"amount\": \"12000000000000000000000000000000\",
+                 \"source\": \"xrb_3dcfozsmekr1tr9skf1oa5wbgmxt81qepfdnt7zicq5x3hk65fg4fqj58mbr\"
             }
         },
         \"xrb_3t6k35gi95xu6tergt6p69ck76ogmitsa8mnijtpxm9fkcm736xtoncuohr3\": {
@@ -487,9 +491,46 @@ RSpec.describe Nanook::Wallet do
     )
 
     response = Nanook.new.wallet(wallet_id).pending(detailed: true)
-    binding.pry
     expect(response.keys).to have(2).items
     expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"]).to have(2).items
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"][0][:amount]).to eq 6
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"][0][:source]).to eq "xrb_3dcfozsmekr1tr9skf1oa5wbgmxt81qepfdnt7zicq5x3hk65fg4fqj58mbr"
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"][0][:block]).to eq "142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D"
+  end
+
+  it "wallet pending with detailed and unit raw" do
+    stub_request(:post, uri).with(
+      body: "{\"action\":\"wallet_pending\",\"wallet\":\"#{wallet_id}\",\"count\":\"1000\",\"source\":\"true\"}",
+      headers: headers
+    ).to_return(
+      status: 200,
+      body: "{\"blocks\": {
+        \"xrb_1111111111111111111111111111111111111111111111111117353trpda\": {
+            \"142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\": {
+                 \"amount\": \"6000000000000000000000000000000\",
+                 \"source\": \"xrb_3dcfozsmekr1tr9skf1oa5wbgmxt81qepfdnt7zicq5x3hk65fg4fqj58mbr\"
+            },
+            \"242A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D\": {
+                 \"amount\": \"12000000000000000000000000000000\",
+                 \"source\": \"xrb_3dcfozsmekr1tr9skf1oa5wbgmxt81qepfdnt7zicq5x3hk65fg4fqj58mbr\"
+            }
+        },
+        \"xrb_3t6k35gi95xu6tergt6p69ck76ogmitsa8mnijtpxm9fkcm736xtoncuohr3\": {
+            \"4C1FEEF0BEA7F50BE35489A1233FE002B212DEA554B55B1B470D78BD8F210C74\": {
+                 \"amount\": \"106370018000000000000000000000000\",
+                 \"source\": \"xrb_13ezf4od79h1tgj9aiu4djzcmmguendtjfuhwfukhuucboua8cpoihmh8byo\"
+            }
+        }
+      }}",
+      headers: {}
+    )
+
+    response = Nanook.new.wallet(wallet_id).pending(detailed: true, unit: :raw)
+    expect(response.keys).to have(2).items
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"]).to have(2).items
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"][0][:amount]).to eq 6000000000000000000000000000000
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"][0][:source]).to eq "xrb_3dcfozsmekr1tr9skf1oa5wbgmxt81qepfdnt7zicq5x3hk65fg4fqj58mbr"
+    expect(response["xrb_1111111111111111111111111111111111111111111111111117353trpda"][0][:block]).to eq "142A538F36833D1CC78B94E11C766F75818F8B940771335C6C1B8AB880C5BB1D"
   end
 
 end